+++ tests/Tests/Utils/Classes/Something.php
@@ -2,6 +2,6 @@
 
 declare(strict_types=1);
 
-namespace MagicPush\CliToolkit\Tests\Tests\ToolBelt\Classes;
+namespace MagicPush\CliToolkit\Tests\Tests\Utils\Classes;
 
 abstract class Something { }