+++ tests/Tests/Utils/UtilsMock.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace MagicPush\CliToolkit\Tests\Tests\ToolBelt;
+namespace MagicPush\CliToolkit\Tests\Tests\Utils;
 
-use MagicPush\CliToolkit\ToolBelt;
+use MagicPush\CliToolkit\Utils;
 
 /**
  * Contains logic needed for tests only.
  */
-final class ToolBeltMock extends ToolBelt {
+final class UtilsMock extends Utils {
     public static function _getCachedTopmostProjectRootDirectory(): ?string {
         return static::$topmostProjectRootDirectory ?? null;
     }